requirements.txt
asgiref==3.2.10
Django==3.0.7
django-bootstrap-pagination==1.7.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-rest==0.8.6
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
